% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareOrig.R
\name{compareOrig}
\alias{compareOrig}
\title{function to compare the original matrix of correct classes to
each component of the output object for a certain classifier}
\usage{
compareOrig(nlcvObj, techn)
}
\arguments{
\item{nlcvObj}{return of the \code{\link{nlcv}} function}

\item{techn}{technique for which the comparison to correct classes should be made}
}
\value{
list with for each number of features selected,
a matrix of logical values indicating whether the classifier results
correspond (TRUE) or not (FALSE) to the original values to be classified
}
\description{
function to compare the original matrix of correct classes to
each component of the output object for a certain classifier
}
