% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nlcvRF_R}
\alias{nlcvRF_R}
\title{nlcv results on random data with random forest feature selection}
\format{\code{nlcv} object}
\usage{
nlcvRF_R
}
\description{
This data set contains the \link{nlcv} results of 
selection of features with random forest on a randomly generated dataset.
}
