% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nlcvRF_SHS}
\alias{nlcvRF_SHS}
\title{nlcv results on strong hetero signal data with random forest feature selection}
\format{
\code{nlcv} object
}
\usage{
data(nlcvRF_SHS)
}
\description{
This data set contains the \link{nlcv} results of 
selection of features with random forest on a dataset with strong hetero signal.
}
