% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nlcvRF_SS}
\alias{nlcvRF_SS}
\title{nlcv results on strong signal data a with random forest feature selection}
\format{\code{nlcv} object}
\usage{
nlcvRF_SS
}
\description{
This data set contains the \link{nlcv} results of 
selection of features with random forest on a dataset with strong signal.
}
