% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nlcvRF_WHS}
\alias{nlcvRF_WHS}
\title{nlcv results on weak signal data with random forest feature selection}
\format{
\code{nlcv} object
}
\usage{
data(nlcvRF_WHS)
}
\description{
This data set contains the \link{nlcv} results of 
selection of features with random forest on a weak signal dataset.
}
