% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nlcvRF_WS}
\alias{nlcvRF_WS}
\title{nlcv results on weak hetero signal data with random forest feature selection}
\format{\code{nlcv} object}
\usage{
nlcvRF_WS
}
\description{
This data set contains the \link{nlcv} results of 
selection of features with random forest on a weak hetero signal dataset.
}
