% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nlcvTT_R}
\alias{nlcvTT_R}
\title{nlcv results on random data with t-test feature selection}
\format{
\code{nlcv} object
}
\usage{
data(nlcvTT_R)
}
\description{
This data set contains the \link{nlcv} results of 
selection of features with t-test on a randomly generated dataset.
}
