% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nlcvTT_SHS}
\alias{nlcvTT_SHS}
\title{nlcv results on strong hetero signal data with t-test feature selection}
\format{\code{nlcv} object}
\usage{
nlcvTT_SHS
}
\description{
This data set contains the \link{nlcv} results of 
selection of features with t-test on a dataset with strong hetero signal.
}
