% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nlcvTT_SS}
\alias{nlcvTT_SS}
\title{nlcv results on strong signal data a with t-test feature selection}
\format{
\code{nlcv} object
}
\usage{
data(nlcvTT_SS)
}
\description{
This data set contains the \link{nlcv} results of 
selection of features with t-test on a dataset with strong signal.
}
