% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nlcvTT_WHS}
\alias{nlcvTT_WHS}
\title{nlcv results on weak signal data with t-test feature selection}
\format{\code{nlcv} object}
\usage{
nlcvTT_WHS
}
\description{
This data set contains the \link{nlcv} results of 
selection of features with t-test on a weak signal dataset.
}
