% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nlcvTT_WS}
\alias{nlcvTT_WS}
\title{nlcv results on weak hetero signal data with t-test feature selection}
\format{
\code{nlcv} object
}
\usage{
data(nlcvTT_WS)
}
\description{
This data set contains the \link{nlcv} results of 
selection of features with t-test on a weak hetero signal dataset.
}
