\name{nldaI}
\alias{nldaI}
\docType{data}
\title{new MLInterfaces schema for lda from MASS}
\description{
  This interface keeps track of the predictions on
  the training and test set, contrary to the ldaI
  interface that is made available in the MLInterfaces
  package.
}
\usage{nldaI}
\details{
  nldaI is an object of class 'learnerSchema' and can be
  used as such in calls to MLearn (from MLInterfaces).
}
\seealso{See Also \code{\link[MLInterfaces]{ldaI}}}
\keyword{manip}
