% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pamrI.R
\name{pamrIconverter}
\alias{pamrIconverter}
\title{convert from \code{pamrML} to \code{classifierOutput}}
\usage{
pamrIconverter(obj, data, trainInd)
}
\arguments{
\item{obj}{object as returned by pamrML i.e. of class \code{pamrML}}

\item{data}{original data used as input for MLearn}

\item{trainInd}{training indices used as input to MLearn}
}
\value{
object of class \code{classifierOutput}
}
\description{
convert from \code{pamrML} to \code{classifierOutput}
}
