% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pamrI.R
\name{predict.pamrML}
\alias{predict.pamrML}
\title{predict \code{pamrML} object}
\usage{
\method{predict}{pamrML}(object, newdata, ...)
}
\arguments{
\item{object}{\code{pamrML} object}

\item{newdata}{new data}

\item{...}{additional parameters for the \link{pamr.predict} function}
}
\value{
output of the \code{pamr.predict} function
}
\description{
predict \code{pamrML} object
}
