% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confusionMatrix.R
\name{print.nlcvConfusionMatrix}
\alias{print.nlcvConfusionMatrix}
\title{print object \code{nlcvConfusionMatrix}}
\usage{
\method{print}{nlcvConfusionMatrix}(x, ...)
}
\arguments{
\item{x}{object of class \code{nlcvConfusionMatrix}}

\item{...}{additional parameters for the \code{print} function}
}
\value{
no returned value, the object is printed in the output
}
\description{
print object \code{nlcvConfusionMatrix}
}
