% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pamrI.R
\name{print.pamrML}
\alias{print.pamrML}
\title{print \code{pamrML} object}
\usage{
\method{print}{pamrML}(x, ...)
}
\arguments{
\item{x}{object of class \code{pamrML}}

\item{\dots}{additional parameters for the \code{print} function}
}
\description{
print \code{pamrML} object
}
