% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcrPlot.R
\name{print.summary.mcrPlot}
\alias{print.summary.mcrPlot}
\title{\code{print} function for \code{summary.mcrPlot} object}
\usage{
\method{print}{summary.mcrPlot}(x, digits = 2, ...)
}
\arguments{
\item{x}{Object of class 'summary.mcrPlot' as produced by the function of
the same name}

\item{digits}{number of digits to be passed to the default print method}

\item{\dots}{additional parameters for the \code{print.default} function}
}
\description{
\code{print} function for \code{summary.mcrPlot} object
}
