% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcrPlot.R
\name{summary.mcrPlot}
\alias{summary.mcrPlot}
\title{\code{summary} function for \code{mcrPlot} object}
\usage{
\method{summary}{mcrPlot}(object, ...)
}
\arguments{
\item{object}{Object of class 'mcrPlot' as produced by the function of the
same name}

\item{...}{additional arguments, not used here}
}
\description{
\code{summary} function for \code{mcrPlot} object
}
