% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limma.R
\name{limmaTwoGroups}
\alias{limmaTwoGroups}
\title{Wrapper around limma for the comparison of two groups}
\usage{
limmaTwoGroups(object, group)
}
\arguments{
\item{object}{object of class ExpressionSet}

\item{group}{string indicating the variable defining the two groups to be
compared}
}
\value{
\code{topTable} output for the second (i.e. slope) coefficient of
the linear model.
}
\description{
Wrapper around limma for the comparison of two groups
}
\details{
Basically, the wrapper combines the \code{lmFit}, \code{eBayes} and
\code{topTable} steps
}
\references{
Smyth, G. K. (2004). Linear models and empirical Bayes methods
for assessing differential expression in microarray experiments.
\emph{Statistical Applications in Genetics and Molecular Biology}, Vol. 3,
No. 1, Article 3.
\doi{doi:10.2202/1544-6115.1027}
}
\author{
Tobias Verbeke
}
\keyword{models}
\keyword{regression}
