% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nldaI.R
\docType{data}
\name{nldaI}
\alias{nldaI}
\title{new MLInterfaces schema for lda from MASS}
\format{
An object of class \code{learnerSchema} of length 1.
}
\usage{
nldaI
}
\description{
This interface keeps track of the predictions on the training and test set,
contrary to the ldaI interface that is made available in the MLInterfaces
package.
}
\details{
nldaI is an object of class 'learnerSchema' and can be used as such in calls
to MLearn (from MLInterfaces).
}
\seealso{
See Also \code{\link[MLInterfaces]{ldaI}}
}
\keyword{manip}
